+++ ChangeLog.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## Version [0.5.0](https://github.com/MC2IT/Agicap.net/compare/v0.4.0...v0.5.0)
+- Added the `NestedList` class.
 - Added the `Payments.SynchronizedBeneficiary` class.
+- Added the `BeneficiarySynchronizationErrorCode` enumeration.
 
 ## Version [0.4.0](https://github.com/MC2IT/Agicap.net/compare/v0.3.0...v0.4.0)
 - Added the `HttpResponseException` class.